requirements.txt
Flask==1.1.2
MarkupSafe==1.1.1
requests==2.24.0
gunicorn==20.0.4
psycopg2-binary==2.8.5
cloudscraper==1.2.40
pytz==2020.1
beautifulsoup4==4.9.1
